requirements.txt
fastapi
uvicorn 
celery
redis
httpx
python-dotenv
